refactor: use ColumnFamily.count() for getIncidentCount()
Agent-Logs-Url: https://github.com/camunda/camunda/sessions/d5d1d44d-da6a-45cb-b8d7-ddc77c7884a3

+++ zeebe/engine/src/main/java/io/camunda/zeebe/engine/state/immutable/IncidentState.java
@@ -29,5 +29,5 @@ public interface IncidentState {
       long elementInstanceKey, ObjLongConsumer<IncidentRecord> resolver);
 
   /** Returns the total number of incidents currently stored in the state. */
-  int getIncidentCount();
+  long getIncidentCount();
 }
